:recycle: Refactor product
