+++ src/controllers/product/product_controller.py
@@ -1,4 +1,4 @@
-from src.main.domain.entities.products_entity import Product
+from src.main.domain.models.models import Products
 from src.database.database import get_connect
 from typing import Dict
 
@@ -8,7 +8,7 @@ class ProductCreatorController:
     def create(self, product: dict) -> Dict:
         try:
             db = get_connect()
-            new_product = Product.create(
+            new_product = Products.create(
                 name=product.get('name'),
                 price=int(product.get('price') * 1000),
                 description=product.get('description'),
@@ -27,3 +27,20 @@ class ProductCreatorController:
                 "id": id_created_product
             }
         }
+
+
+class ProductGetController:
+
+    @staticmethod
+    def get_all() -> Dict:
+
+        try:
+            with get_connect():
+                products = Products.select()
+                print(products)
+                return {
+                    "products": [{**product.serialize(), "price": "{:.2f}".format(product.price / 1000)} for product in products]
+                }
+        except Exception as e:
+            print(f"Error: {e}")
+            return {}
