+++ src/views/product_view/product_view.py
@@ -1,13 +1,12 @@
 from src.views.http_types.http_request import HttpRequest
 from src.views.http_types.http_response import HttpResponse
-from src.controllers.product.product_creator_controller import ProductCreatorController
-from src.controllers.product.product_get_controller import ProductGetController
+from src.controllers.product.product_controller import Products
 
 
 class ProductView:
 
     def validate_and_create_product(self, http_request: HttpRequest) -> HttpResponse:
-        product_creator_controller = ProductCreatorController()
+        product_creator_controller = Products()
         body = http_request.body
 
         result = product_creator_controller.create(body)
@@ -17,7 +16,7 @@ class ProductView:
 
 
     def find_all_product(self) -> HttpResponse:
-        product_get_controller = ProductGetController()
+        product_get_controller = Products()
 
         result = product_get_controller.get_all()
